requirements.txt
atlassian-python-api
dotenv